buildercore.core, hostname_struct moved to buildercore.cfngen adds a warning about some potentially unsafe usage of project data
