+++ projects/elife.yaml
@@ -49,10 +49,11 @@ defaults:
             # lsh@2022-07: all instances run on type T3 now, this option should be removed.
             cpu-credits: standard
             # whether the first EC2 node should get an external DNS entry such as env--project--1.elifesciences.org
-            # only makes sense if cluster-size > 1
+            # only makes sense if cluster-size > 1 with no load balancer.
+            # lsh@2022-09: unused feature, not sure which project it was built for.
             dns-external-primary: false
             # whether the EC2 nodes should get a per-node internal DNS entry such as env--project--1.elife.internal
-            # only makes sense if cluster-size > 1
+            # only makes sense if cluster-size > 1 with no load balancer (see 'search' project).
             dns-internal: false
             # override 'ext' (only supported key)
             # for some EC2 instances
